Added for.prev and for.next

+++ include/symbols/types/primitives/LgsShort.h
@@ -3,7 +3,7 @@
 
 class LgsShort final : public LgsType {
 public:
-    static constexpr auto name = "Int";
+    static constexpr auto name = "Short";
 
     LgsShort() {
         isPrimitive = true;
